+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[2.1.17](https://github.com/ElrondNetwork/dapp-core/pull/491)] - 2022-12-01
+- [Fixed setting incremental nonces when signing with ledger](https://github.com/ElrondNetwork/dapp-core/pull/492)
+
 ## [[2.1.16](https://github.com/ElrondNetwork/dapp-core/pull/490)] - 2022-11-29
 - [Fixed logging out with wallet provider](https://github.com/ElrondNetwork/dapp-core/pull/489)
 - [Added `status` param to `getTransactions` API](https://github.com/ElrondNetwork/dapp-core/pull/488)